# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Chop Plan API - prepaid lunch subscription service for Lagos restaurants
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendor/profile:
    get:
      operationId: getVendorProfile
      tags: [vendor]
      summary: Get logged-in vendor profile
      responses:
        "200":
          description: Vendor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorProfile"
    patch:
      operationId: updateVendorProfile
      tags: [vendor]
      summary: Update vendor profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorProfile"
...
  /vendor/meals:
    get:
      operationId: listMyMeals
      tags: [vendor]
      summary: Get vendor's own meal listings
      responses:
        "200":
          description: Meal list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meal"
    post:
      operationId: createMeal
      tags: [vendor]
      summary: Add a new meal to the menu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
  /vendor/meals/{mealId}:
    patch:
      operationId: updateMeal
      tags: [vendor]
      summary: Update a meal
      parameters:
        - name: mealId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
    delete:
      operationId: deleteMeal
      tags: [vendor]
      summary: Delete a meal
      parameters:
        - name: mealId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /vendor/customers/{subscriptionId}/schedule:
    get:
      operationId: getVendorCustomerSchedule
      tags: [vendor]
      summary: Get the daily pickup schedule for one of the vendor's subscribers
      parameters:
        - name: subscriptionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Schedule days
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleDay"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /vendor/wallet/withdraw:
    post:
      operationId: withdrawFromWallet
      tags: [vendor]
      summary: Withdraw available funds from the wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawInput"
      responses:
        "200":
          description: Updated wallet summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorWallet"
        "400":
          description: Invalid withdrawal amount
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── ADMIN ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Auth ──────────────────────────────────────────────────────────────
...
    Meal:
      type: object
      required: [id, name, description, priceNaira, imageUrl, available]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        priceNaira:
          type: integer
        imageUrl:
          type: string
        available:
          type: boolean
        category:
          type: ["string", "null"]

    # ── User ─────────────────────────────────────────────────────────────
...
    VendorProfile:
      type: object
      required: [id, businessName, ownerName, email, phone, area, cuisineType]
      properties:
        id:
          type: integer
        businessName:
          type: string
        ownerName:
          type: string
        email:
          type: string
        phone:
          type: string
        area:
          type: string
        cuisineType:
          type: string
        description:
          type: ["string", "null"]
        coverImage:
          type: ["string", "null"]
    VendorProfileUpdate:
      type: object
      properties:
        businessName:
          type: string
        ownerName:
          type: string
        phone:
          type: string
        area:
          type: string
        cuisineType:
          type: string
        description:
          type: string
...
    MealInput:
      type: object
      required: [name, description, priceNaira, imageUrl]
      properties:
        name:
          type: string
        description:
          type: string
        priceNaira:
          type: integer
        imageUrl:
          type: string
        category:
          type: string
        available:
          type: boolean
    MealUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        priceNaira:
          type: integer
        imageUrl:
          type: string
        category:
          type: string
        available:
          type: boolean

    # ── Order tracking & wallet ─────────────────────────────────────────────
    ScheduleDay:
      type: object
      required: [id, dayNumber, scheduledDate, status, confirmedAt]
      properties:
        id:
          type: integer
        dayNumber:
          type: integer
        scheduledDate:
          type: string
          format: date
        status:
          type: string
          enum: [pending, confirmed]
        confirmedAt:
          type: ["string", "null"]
          format: date-time
    WalletWithdrawal:
      type: object
      required: [id, amountNaira, createdAt]
      properties:
        id:
          type: integer
        amountNaira:
          type: integer
        createdAt:
          type: string
          format: date-time
    VendorWallet:
      type: object
      required: [earnedNaira, withdrawableNaira, withdrawnNaira, withdrawals]
      properties:
        earnedNaira:
          type: integer
          description: Total earned to date from confirmed pickups
        withdrawableNaira:
          type: integer
          description: Amount currently available to withdraw
        withdrawnNaira:
          type: integer
        withdrawals:
          type: array
          items:
            $ref: "#/components/schemas/WalletWithdrawal"
    WithdrawInput:
      type: object
      required: [amountNaira]
      properties:
        amountNaira:
          type: integer
          minimum: 1

    # ── Admin ────────────────────────────────────────────────────────────────
...