# lib/api-zod/src/generated/types/vendorWallet.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Chop Plan API - prepaid lunch subscription service for Lagos restaurants
 * OpenAPI spec version: 0.1.0
 */
import type { WalletWithdrawal } from './walletWithdrawal';

export interface VendorWallet {
  /** Total earned to date from confirmed pickups */
  earnedNaira: number;
  /** Amount currently available to withdraw */
  withdrawableNaira: number;
  withdrawnNaira: number;
  withdrawals: WalletWithdrawal[];
}