# lib/api-zod/src/generated/types/withdrawInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Chop Plan API - prepaid lunch subscription service for Lagos restaurants
 * OpenAPI spec version: 0.1.0
 */

export interface WithdrawInput {
  /** @minimum 1 */
  amountNaira: number;
}